python-services/requirements.txt
flask
flask-cors
sentence-transformers
numpy
python-dotenv
psycopg2-binary